chore: CombineExtension -> MVMCombineExtension 이름수정
