--- iOS-MyCarMaster/CombineExtension/CombineExtension/CombineExtension.h
@@ -1,17 +1,0 @@
-//
-//  CombineExtension.h
-//  CombineExtension
-//
-//  Created by SEUNGMIN OH on 2023/08/07.
-//
-
-#import <Foundation/Foundation.h>
-
-//! Project version number for CombineExtension.
-FOUNDATION_EXPORT double CombineExtensionVersionNumber;
-
-//! Project version string for CombineExtension.
-FOUNDATION_EXPORT const unsigned char CombineExtensionVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <CombineExtension/PublicHeader.h>
-
