+++ iOS-MyCarMaster/MCMCombineExtension/MCMCombineExtension/Common/Sink.swift
